test: split the pty-delivery suite under the file-size cap
Adding the ready-engine fakes pushed pty-delivery.test.ts to 503 lines. The
pure engine-key resolver cases move to pty-engine-key.test.ts; all 29 tests
are preserved.
